update docs, scripts, version, openapi info
