+++ .vscode/settings.json
@@ -167,6 +167,7 @@
         "mccabe",
         "metaclass",
         "mkdocs",
+        "mkdocsoad",
         "mkdocstrings",
         "mktime",
         "mogami",
@@ -178,6 +179,7 @@
         "mtxr",
         "multiarch",
         "mypy",
+        "neoteroi",
         "netfilter",
         "ninoseki",
         "noconfirm",
