+++ Makefile
@@ -25,6 +25,7 @@ help:
 	@echo " forlint              Alias for 'make format && make lint'"
 	@echo " requirements         Generate requirements/*.txt files"
 	@echo " test                 Run the tests"
+	@echo " openapi              Generate the OpenAPI documentation"
 	@echo " docs                 Generate the documentation"
 	@echo " docs-live            Generate the documentation in 'live' mode"
 	@echo " clean                Remove unneeded files (__pycache__, .mypy_cache, etc.)"
@@ -66,8 +67,12 @@ test:
 	$(PYTHON) scripts/test.py
 	@echo "html report: file://`pwd`/${.REPORTS_DIR}/html/index.html"
 
+.PHONY: openapi
+openapi:
+	$(PYTHON) scripts/openapi.py
+
 .PHONY: .before-docs
-.before-docs:
+.before-docs: openapi
 	$(PYTHON) scripts/docs.py before
 
 .PHONY: .after-docs
@@ -75,7 +80,7 @@ test:
 	$(PYTHON) scripts/docs.py after
 
 .PHONY: docs
-docs:
+docs: openapi
 	$(PYTHON) scripts/docs.py
 	$(PYTHON) scripts/docs.py after
