+++ docs/reference/openapi.md
@@ -1,0 +1,1 @@
+[OAD(./docs/openapi.json)]
