+++ pyproject.toml
@@ -144,7 +144,8 @@ docs = [
   'mkdocs-mermaid2-plugin==1.2.1',
   'mkdocs-minify-html-plugin==0.3.1',
   'mkdocstrings[crystal,python]==0.29.1',
-  'mkdocstrings-python==1.16.10'
+  'mkdocstrings-python==1.16.10',
+  'neoteroi-mkdocs==1.1.0',
 ]
 
 [tool.hatch.version]
