+++ requirements/docs.txt
@@ -9,3 +9,4 @@ mkdocs-minify-html-plugin==0.3.1
 mkdocs==1.6.1
 mkdocstrings-python==1.16.10
 mkdocstrings[crystal,python]==0.29.1
+neoteroi-mkdocs==1.1.0
