+++ waldiez_runner/_version.py
@@ -2,4 +2,4 @@
 # Copyright (c) 2024 - 2025 Waldiez and contributors.
 """Version information for waldiez_runner."""
 
-__version__ = "0.3.12"
+__version__ = "0.4.4"
